H3HexagonLayer: Upgrade to use deck.gl's H3HexagonLayer
